feat: complete package refactoring cleanup - fix imports and KDoc references
- Phase 1: Verified all duplicate files already deleted
- Phase 2: Fixed 28 broken imports in 6 source files
  - DslContainerRegistry.kt, DslNavigationConfig.kt, DslScopeRegistry.kt
  - DslScreenRegistry.kt, DslTransitionRegistry.kt, CompositeNavigationConfig.kt
- Phase 3: Verified test imports already fixed
- Phase 4: Updated KDoc references in 13 files to new package paths
  - navigation.tree.* → navigation.internal.tree.*
  - dsl.registry.* → registry.*
  - compose.wrapper.* → compose.scope.*
  - compose.render.* → compose.scope.*/compose.internal.*
  - compose.navback.* → compose.internal.navback.*
- Phase 5: Verified all builds and tests pass
  - quo-vadis-core:build ✓
  - quo-vadis-core:allTests ✓
  - quo-vadis-ksp:build ✓
  - All modules compile ✓
  - Updated detekt baseline
