+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/dsl/DslScopeRegistry.kt
@@ -1,8 +1,8 @@
 package com.jermey.quo.vadis.core.dsl
 
 import com.jermey.quo.vadis.core.InternalQuoVadisApi
-import com.jermey.quo.vadis.core.dsl.registry.ScopeRegistry
-import com.jermey.quo.vadis.core.navigation.NavDestination
+import com.jermey.quo.vadis.core.registry.ScopeRegistry
+import com.jermey.quo.vadis.core.navigation.destination.NavDestination
 import kotlin.reflect.KClass
 
 /**
