--- quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/dsl/registry/CompositeTransitionRegistry.kt
@@ -1,19 +1,0 @@
-package com.jermey.quo.vadis.core.dsl.registry
-
-import com.jermey.quo.vadis.core.InternalQuoVadisApi
-import com.jermey.quo.vadis.core.compose.animation.NavTransition
-import kotlin.reflect.KClass
-
-/**
- * Composite transition registry that delegates to secondary first, then primary.
- */
-@InternalQuoVadisApi
-internal class CompositeTransitionRegistry(
-    private val primary: TransitionRegistry,
-    private val secondary: TransitionRegistry
-) : TransitionRegistry {
-
-    override fun getTransition(destinationClass: KClass<*>): NavTransition? {
-        return secondary.getTransition(destinationClass) ?: primary.getTransition(destinationClass)
-    }
-}
