+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/navigation/config/CompositeNavigationConfig.kt
@@ -3,20 +3,21 @@
 package com.jermey.quo.vadis.core.navigation.config
 
 import com.jermey.quo.vadis.core.InternalQuoVadisApi
-import com.jermey.quo.vadis.core.navigation.NavNode
-import com.jermey.quo.vadis.core.navigation.Navigator
-import com.jermey.quo.vadis.core.dsl.registry.CompositeContainerRegistry
-import com.jermey.quo.vadis.core.dsl.registry.CompositeScopeRegistry
-import com.jermey.quo.vadis.core.dsl.registry.CompositeScreenRegistry
-import com.jermey.quo.vadis.core.dsl.registry.CompositeTransitionRegistry
-import com.jermey.quo.vadis.core.dsl.registry.ContainerRegistry
-import com.jermey.quo.vadis.core.dsl.registry.PaneRoleRegistry
-import com.jermey.quo.vadis.core.dsl.registry.ScopeRegistry
-import com.jermey.quo.vadis.core.dsl.registry.ScreenRegistry
-import com.jermey.quo.vadis.core.dsl.registry.TransitionRegistry
-import com.jermey.quo.vadis.core.navigation.DeepLink
-import com.jermey.quo.vadis.core.dsl.registry.DeepLinkRegistry
-import com.jermey.quo.vadis.core.navigation.NavDestination
+import com.jermey.quo.vadis.core.navigation.node.NavNode
+import com.jermey.quo.vadis.core.navigation.navigator.Navigator
+import com.jermey.quo.vadis.core.registry.internal.CompositeContainerRegistry
+import com.jermey.quo.vadis.core.registry.internal.CompositeScopeRegistry
+import com.jermey.quo.vadis.core.registry.internal.CompositeScreenRegistry
+import com.jermey.quo.vadis.core.registry.internal.CompositeTransitionRegistry
+import com.jermey.quo.vadis.core.registry.ContainerRegistry
+import com.jermey.quo.vadis.core.registry.PaneRoleRegistry
+import com.jermey.quo.vadis.core.registry.ScopeRegistry
+import com.jermey.quo.vadis.core.registry.ScreenRegistry
+import com.jermey.quo.vadis.core.registry.TransitionRegistry
+import com.jermey.quo.vadis.core.navigation.destination.DeepLink
+import com.jermey.quo.vadis.core.registry.DeepLinkRegistry
+import com.jermey.quo.vadis.core.navigation.destination.NavDestination
+import com.jermey.quo.vadis.core.navigation.internal.config.EmptyNavigationConfig
 import com.jermey.quo.vadis.core.navigation.pane.PaneRole
 import kotlin.reflect.KClass
