+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/navigation/destination/DeepLinkResult.kt
@@ -32,7 +32,7 @@ package com.jermey.quo.vadis.core.navigation.destination
  * }
  * ```
  *
- * @see com.jermey.quo.vadis.core.dsl.registry.CompositeDeepLinkRegistry
+ * @see com.jermey.quo.vadis.core.registry.internal.CompositeDeepLinkRegistry
  */
 public sealed class DeepLinkResult {
     /**
