+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/navigation/node/TabNode.kt
@@ -34,8 +34,8 @@ import kotlin.uuid.Uuid
  *
  * ## Scope-Aware Navigation
  *
- * When [scopeKey] is set, [com.jermey.quo.vadis.core.navigation.tree.TreeMutator.push] with a
- * [com.jermey.quo.vadis.core.dsl.registry.ScopeRegistry] will check
+ * When [scopeKey] is set, [com.jermey.quo.vadis.core.navigation.internal.tree.TreeMutator.push] with a
+ * [com.jermey.quo.vadis.core.registry.ScopeRegistry] will check
  * if destinations belong to this container's scope. Out-of-scope destinations
  * navigate to the parent stack instead, preserving the tab container for
  * predictive back gestures.
@@ -50,7 +50,7 @@ import kotlin.uuid.Uuid
  * @property parentKey Key of the containing node (null if root)
  * @property stacks List of StackNodes, one per tab
  * @property activeStackIndex Index of the currently active tab (0-based)
- * @property wrapperKey Key used to lookup the wrapper in [com.jermey.quo.vadis.core.dsl.registry.ContainerRegistry].
+ * @property wrapperKey Key used to lookup the wrapper in [com.jermey.quo.vadis.core.registry.ContainerRegistry].
  *   This is typically the simple name of the tab class (e.g., "MainTabs")
  *   and is used by the hierarchical renderer to find the correct wrapper.
  *   Defaults to null, which means no custom wrapper is registered.
